requirements.txt
asgiref==3.8.1
confluent-kafka==2.4.0
Django==5.0.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests~=2.32.3
sqlparse==0.5.0